+++ packages/terra-worklist-data-grid/src/proptypes/cellShape.js
@@ -2,10 +2,9 @@ import PropTypes from 'prop-types';
 
 const cellShape = PropTypes.shape({
   /**
-   * Content that will rendered within the Cell.
+   * Content that will be rendered within the Cell.
    */
   content: PropTypes.node,
-
   /**
    *  Boolean indicating if cell contents are masked.
    */