+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+Added
+ * Added additional keyboard navigation in `terra-worklist-data-grid`.
+ * Added support for cell selection.
 
 ## 0.2.0 - (July 4, 2023)
 
@@ -12,7 +15,6 @@
   * Added hover state for selectable cells.
   * Added blank cell support.
   * Added ability to allow consumers to specify the sort indicator and error state for columns.
-  * Added documentation and tests for additional keyboard navigation in `terra-worklist-data-grid`.
 
 ## 0.1
   * Initial stable release.